cleanup playlist stuff
